api/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-multipart
watchfiles
python-dotenv
passlib[bcrypt]
